all logic and screens are in the package folder
