+++ packages/polindrome/lib/domain/repository/base.dart
@@ -1,0 +1,3 @@
+abstract class Polindrome{
+  bool checkPalindrome(String inputString);
+}
